Remove unused ACL related functions
