+++ backup/queries_acl.go
@@ -126,7 +126,7 @@ func GetMetadataForObjectType(connectionPool *dbconn.DBConn, params MetadataQuer
 		aclStr = fmt.Sprintf(`CASE
 		WHEN %[1]s IS NULL THEN NULL
 		WHEN array_upper(%[1]s, 1) = 0 THEN %[1]s[0]
-		ELSE UNNEST(%[1]s) END`, params.ACLField)
+		ELSE unnest(%[1]s) END`, params.ACLField)
 		kindStr = fmt.Sprintf(`CASE
 		WHEN %[1]s IS NULL THEN ''
 		WHEN array_upper(%[1]s, 1) = 0 THEN 'Empty'
@@ -150,7 +150,7 @@ WHERE %s`, params.SchemaField, SchemaFilterClause("n"))
 	secCols := ""
 	secStr := ""
 	if connectionPool.Version.AtLeast("6") {
-		secCols = "COALESCE(sec.label,'') AS securitylabel, COALESCE(sec.provider, '') AS securitylabelprovider,"
+		secCols = "coalesce(sec.label,'') AS securitylabel, coalesce(sec.provider, '') AS securitylabelprovider,"
 		secTable := "pg_seclabel"
 		secSubidStr := " AND sec.objsubid = 0"
 		if params.Shared {
